TLS-Core/src/main/java/de/rub/nds/tlsattacker/core/protocol/handler/extension/TokenBindingExtensionHandler.java
/*
 * TLS-Attacker - A Modular Penetration Testing Framework for TLS
 *
 * Copyright 2014-2023 Ruhr University Bochum, Paderborn University, Technology Innovation Institute, and Hackmanit GmbH
 *
 * Licensed under Apache License, Version 2.0
 * http://www.apache.org/licenses/LICENSE-2.0.txt
 */
package de.rub.nds.tlsattacker.core.protocol.handler.extension;

import de.rub.nds.tlsattacker.core.constants.TokenBindingKeyParameters;
import de.rub.nds.tlsattacker.core.constants.TokenBindingVersion;
import de.rub.nds.tlsattacker.core.layer.context.TlsContext;
import de.rub.nds.tlsattacker.core.protocol.message.extension.TokenBindingExtensionMessage;
import java.util.ArrayList;

public class TokenBindingExtensionHandler extends ExtensionHandler<TokenBindingExtensionMessage> {

    public TokenBindingExtensionHandler(TlsContext tlsContext) {
        super(tlsContext);
    }

    @Override
    public void adjustTLSExtensionContext(TokenBindingExtensionMessage message) {
        tlsContext.setTokenBindingVersion(
                TokenBindingVersion.getExtensionType(message.getTokenBindingVersion().getValue()));
        ArrayList<TokenBindingKeyParameters> tokenbindingKeyParameters = new ArrayList<>();
        for (byte kp : message.getTokenBindingKeyParameters().getValue()) {
            tokenbindingKeyParameters.add(
                    TokenBindingKeyParameters.getTokenBindingKeyParameter(kp));
        }
        tlsContext.setTokenBindingKeyParameters(tokenbindingKeyParameters);
        if (tlsContext.getTalkingConnectionEndType()
                == tlsContext.getChooser().getMyConnectionPeer()) {
            tlsContext.setTokenBindingNegotiatedSuccessfully(true);
        }
    }
}
